Fix deadlock in plugin
169ae75cd17fdff4af5b005d846566a690c0d12e introduced the MumbleEventLoop
class, which was supposed to make the processing of Mumble callbacks
asynchronous. However, the loop was keeping the lock when executing
queued callbacks and if such a callback was to use a Mumble API function
while Mumble's main thread was currently executing one of the plugin
callbacks within ACRE, this API call would block. If the active callback
now tried to queue a new event, this would block as well, since the
event queue was still holding the lock for processing that function.
Thus, a deadlock occured.

Given that the timing had to be right for this to happen, the deadlocks
were now much less common than they were before the introduction of the
event queue, but with a bit of patience, they could still be created.

The solution is to simply release the lock while executing the currently
processed function in the event loop. This allows simultaneous queueing
of new events withouot blocking.