Fix crash due to double-join
Joining a thread makes it unjoinable. Attempting to join it again will
throw an exception, which (due to not being handled) will crash Mumble.
This situation would be created by loading, unloading, loading and
unloading ACRE because the used EventLoop instance is a singleton that
will survive the reloads. However, the used thread object will be
invalid after the second load.

The solution is to simply make starting the thread explicit instead of
implicitly starting it as soon as the loop is instantiated.