Do not load AcreSettings in constructor
If the default path baked into the constructor is not the same as one
that is later loaded, the user will end up with two config files! Only
one will be used.

There is no need to load the config file in the constructor, however.
Hardcoded defaults are already set, and users of the class all load
config files explicitly.