Reimplement NamedPipeServer on Linux
This is the most involved change in this changeset.

On Linux, Arma runs under WINE, but Mumble is native. There's no (known
to me) way to do IPC between WINE processes and native processes; WINE
makes it (purposefully?) difficult.

It may be possible to use Windows 10+ Unix socket compatibility (if WINE
supports it), or some other mechanism. But this solution uses a simple
duplex network socket listening on localhost, with a default port of
19141. The Mumble side is responsible for setting up the server, and the
Arma side connects.

ACRE uses Windows named pipes in "message" mode, which means rather than
a stream of bytes, senders send a stream of messages and receivers
receive whole messages, not partial messages.

This is not the case with a TCP socket, and so a simple length encoding
is used. Whenever a sender has a message to send, they first send a four
byte unsigned integer that is the length of the message to come, and the
receiver first reads exactly four bytes, decodes the length of the
message, then reads exactly that many bytes.

The code is written to be tolerant (as in: will not crash, and will
simply drop the connection and wait for the otherside to reconnect) to
network faults and partial sends/receives. However, given that the
client and server are on the same networking stack, conditions are
ideal: sends and receives should match up exactly, and faults should be
impossible.

Since the port chosen could conflict with another server, possibly
another instance of Mumble/Acre/Arma, it is configurable. On the Mumble
side, it can be configured in the ACRE config file. On the Arma side, it
will be specified as a CBA setting and passed in to the extension via
a parameter to callExtension.