Modify arma2ts to use sockets if running in Wine
When running in Wine, ACRE2Arma.dll will mirror the changes to
NamedPipeServer.cpp for Linux; instead of using named pipes, it will
connect to the socket opened on the other side.

A helper function has been added to detect if running under Wine, by
checking for a function Wine defines in ntdll.