Set LC_NUMERIC=C on Linux
The atof function is locale dependent. Specifically, it uses the
locale-specific decimal separator to convert floating-point strings;
that is, in a locale that uses commas as decimal separators, "0,5" will
convert to 0.5, whereas "0.5" will fail, and convert as 0.0. Strings
coming from Arma will always use a period as a decimal separator, this
causes myriad failures.

On Linux, on startup, we now set LC_NUMERIC to C, which has the behavior
we want.

I'm not certain why this isn't a problem on Win32. There may be a path
in Qt/Mumble that sets locale one way on Linux and another on Windows.